Fixed First Pixel Check
The `get_next_pixel_value` function checks to see if the pixel being updated is the first pixel on the strip, since on some Pico boards, such as the RP2040-Zero, there is an onboard RGB status LED and this is being assigned as the first pixel position on strip 1 by the Modxo firmware.
However, this function was not checking the strip number, meaning it was treating the first pixel of every strip as the status LED, rather than just the first strip. This change fixes that.