Tweaks / Enhancements to RGB LED Support (#23)
* Fixed First Pixel Check

The `get_next_pixel_value` function checks to see if the pixel being updated is the first pixel on the strip, since on some Pico boards, such as the RP2040-Zero, there is an onboard RGB status LED and this is being assigned as the first pixel position on strip 1 by the Modxo firmware.
However, this function was not checking the strip number, meaning it was treating the first pixel of every strip as the status LED, rather than just the first strip. This change fixes that.

* Fix typo in `ws2812_set_color` function

Fixed a typo which was preventing the blue color value for the RGB LEDs from being sent correctly

* Added support for setting pixel format for each LED strip independently

Allows for mixing and matching of LED strips with different pixel formats if required.